Main

Documentation

data Manager

Constructors

Manager

Fields

state :: State

pid :: Int

step :: Int

before :: Integer

delta :: Integer

delay :: Integer

Main

loadManager :: Manager

Carrega o mapa 1F1.txt (= 1.txt mas é válido)

updateControlledPlayer :: Key -> Manager -> Manager

Muda apenas o jogador de direção

updateScreen :: Window -> ColorID -> Manager -> Curses ()

getPlayerList :: State -> [Player]

Devolve a lista de jogadores do State

currentTime :: IO Integer

Tempo

updateTime :: Integer -> Manager -> Manager

Atualiza o tempo

resetTimer :: Integer -> Manager -> Manager

Coloca o delta a 0

nextFrame :: Integer -> Manager -> Manager

Avança o jogo frame a frame

loop :: Window -> Manager -> Curses ()

main :: IO ()

Main de testes

Load Game

runGame :: String -> IO ()

Inicia um jogo a partir do nome do mapa .txt

loadManagerRun :: String -> Manager

Cria o manager com o nome do mapa .txt

Create Game

createGame :: Maze -> IO ()

Cria um jogo recebendo um Maze gerado

putPlayersInMap :: Maze -> [Player]

Cria 4 players

loadManagerCreate :: Maze -> Manager

Carrega o manager com o mapa gerado e coloca os 4 players