Tarefa1

Description

Este módulo tem como objetivo criar o Maze

Funções da construção inicial do Maze

generateRandoms :: Int -> Int -> [Int]

Dado uma semente retorna uma lista de n inteiros aleatórios

intToPiece :: Int -> Piece

Converter inteiro para peça

3 == Food Big

0 <= x < 70 = Food Little

70 < x <= 99 = Wall

createCorridor :: Int -> Corridor -> Maze

Coloca no corredores do Maze espaços para colocar as Walls

convertCorridor :: [Int] -> Corridor

Converte uma lista de Int para Corridor

Adiciona Walls

addSideWalls :: Maze -> Maze

Adiciona Wall nos lados do maze

createWall

Arguments

:: Int
Comprimento do Maze

-> Corridor

Adiciona Wall no top e bottom do Maze

Adiciona Tunel

tunnelCorridor :: Corridor -> Corridor

Cria um Corridor com o Tunel

findMiddleMaze

Arguments

:: Int
Altura do Maze

-> Maze

-> Corridor

Encontra o Corridor do meio do Maze e retorna-o

addTunnelsOdd :: Int -> Maze -> Maze

Adiciona o Tunel para um Maze com altura impar

addTunnelsEven :: Int -> Maze -> Maze

Adiciona o Tunel para um Maze com altura par

addTunnels :: Int -> Maze -> Maze

Adiciona o Tunel para qualquer Maze

Ghost House

ghostHouse

Arguments

:: Int
altura do maze

-> Int
comprimento do maze

-> Maze

-> Maze

Cria a ghostHouse

putCorridor

Arguments

:: Corridor

-> Int
inteiro que percorre as posições onde deve estar a ghostHouse

-> Int
posição do corridor onde começa a colocar Empty na ghostHouse

-> Int
posição do corridor onde acaba de colocar Empty na ghostHouse

-> Int
posição do corridor onde acaba a GhostHouse

-> Corridor

Põe o primeiro corridor da GhostHouse

GhostHouse Par

putSndAndRdCorridorEven

Arguments

:: Maze

-> Int
posição onde começa a ghostHouse

-> Maze

Põe o segundo corredor e terceiro da GhostHouse para maze par

ghostHouseEven

Arguments

:: Maze

-> Int
inteiro que identifica o numero do corredor (0 - corredor com entrada da ghostHouse; 1 - corredor a seguir à entrada; 2 - corredor com o fim da ghostHouse)

-> Int
posição no corridor onde começa a ghostHouse

-> Maze

GhostHouse para maze par

emptyForGhostEven

Arguments

:: Maze

-> Int
altura do maze onde começa a colocar Empty

-> Int
comprimento da GhostHouse + 1 para colocar Empty ao longo do corridor

-> Int
altura do maze onde deixa de colocar Empty e dá o caso de paragem

-> Maze

GhostHouse Impar

putSndAndRdCorridorOdd

Arguments

:: Maze

-> Int
posição onde começa a ghostHouse

-> Maze

Põe o segundo corredor e terceiro da GhostHouse para maze impar

ghostHouseOdd

Arguments

:: Maze

-> Int
inteiro que identifica o numero do corredor (0 - corredor com entrada da ghostHouse; 1 - corredor a seguir à entrada; 2 - corredor com o fim da ghostHouse)

-> Int
posição no corridor onde começa a ghostHouse

-> Maze

emptyForGhostOdd

Arguments

:: Maze

-> Int
altura do maze onde começa a colocar Empty

-> Int
comprimento da GhostHouse + 1 para colocar Empty ao longo do corridor

-> Int
altura do maze onde deixa de colocar Empty e dá o caso de paragem

-> Maze

Gera a Maze

generateMaze

Arguments

:: Int
altura do Maze

-> Int
comprimento do Maze

-> Int
semente do Maze

-> Maze

Gera o Maze final