Tarefa1_2021li1g020

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2021/22.

Documentation

verificarCoordenadas :: [(Peca, Coordenadas)] -> Coordenadas -> Bool

O verificarCoordenadas vê se não existem duas peças com as mesmas coordenadas, o que não pode ser possível

verificarPorta :: [(Peca, Coordenadas)] -> Int

O verificarPorta procura as portas todas do mapa e se só existir uma, então o mapa poderá ser válido

obterSuportes :: [(Peca, Coordenadas)] -> [Coordenadas]

O obterSuportes pega em todas as caixas e obtem o as coordenadas do objeto que poderá estar debaixo delas

verificarSuportes :: [(Peca, Coordenadas)] -> [Coordenadas] -> Bool

A verificarSuportes vai verificar a existência de um bloco ou caixa na lista de coordenadas obtida na função anterior

verificarFlutuantes :: [(Peca, Coordenadas)] -> Bool

Utilizando as funções obterSuportes e verificarSuportes vai procurar a existência de caixas flutuantes, que não podem exisitr

obterMaiorX :: [(Peca, Coordenadas)] -> Int -> Int

O obterMaiorX vai procurar a peça com o maior X

obterMaiorY :: [(Peca, Coordenadas)] -> Int -> Int

O obterMaiorY vai procurar a peça com o maior Y

obterEspacos :: Coordenadas -> Int

O obterEspaços vai medir a área desta matriz

verificarVazio :: [(Peca, Coordenadas)] -> Coordenadas -> Bool

O verificarVazio vai ver se existe algum Vazio no mapa inteiro, se não existir o mapa não pode válido

verificarBase :: [(Peca, Coordenadas)] -> Int -> Int -> Int -> Bool

Verifica se o mapa tem uma base composta simplesmente por blocos

validaPotencialMapa :: [(Peca, Coordenadas)] -> Bool

FUNÇÃO FINAL:

Função final na qual juntamos todas as funções que irão ver se a lista de Peças e Coordenadas poderia ser um Mapa