Tarefa2

Description

Introdução

O objetivo deste módulo é, dada uma descrição do estado do jogo e uma jogada do Pacman, determinar o efeito dessa jogada no estado do jogo.

Objetivos

Os objetivos aqui falados serão apenas os que foram alterandos ou adicionados a esta Tarefa de maneira a que a Tarefa 4 e 5 funcionem corretamente.

As alteração principais feitas nesta tarefa foram:

pacEatGhost -> foi adicionada mais uma variável que representa a lista de jogadores antes de ser feita a jogada, de maneira a que evita com que quando o ghost e o pacman choquem, efetivamente o pacman morre ou o ghost morre

play -> adicionada condiçáo que verifica se o Player a ser movido é Ghost ou Pacman

moveAuxG -> executa o movimento do Ghost

Conclusão

Estas adições e alterações fizeram com que a Tarefa 4 e Tarefa 5 funcionassem, de acordo com o pedido.

Funções do movimento do Pacman

againstWall :: Maze -> Player -> Orientation -> Int

Verifica se está contra a parede

findTunnelCoords :: Maze -> [Coords]

Encontra as coordenadas todas de cada entrada do tunel

moveTunnel :: Coords -> [Coords] -> Bool

Verifica se a coordenada está dentro da lista de todas as entradas do túnel

getPlayerCoordsOri

Arguments

:: Maze

-> Player

-> Orientation
orientação da play

-> Orientation
orientação do Player

-> Coords

Retorna as coordenadas para onde o pacman e ghost se movem

Funções Auxiliares

setPlayerOrientation :: Player -> Orientation -> Player

Muda a orientação de um player

Funções de quando o Pacman come

pacEat :: Player -> Maze -> [Player] -> Player

Aumenta os pontos do pacman e muda o seu PacMode

isGhostInThoseCoords :: Coords -> [Player] -> Bool

Verifica se existe um Ghost naquelas coordenadas

howManyGhostsInThoseCoords :: Coords -> [Player] -> Int -> Int

Verifica quantos ghosts estão naquela coordenada

pacEatGhost

Arguments

:: Maze

-> Player
Player que foi movido

-> [Player]

-> [Player]
Lista de Players inalterada

-> [Player]
Lista de Players antes da jogada

-> Bool
Boolean para utilizar isGhostInThoseCoords

-> [Player]

Comer ou não o fantasma dando a lista de jogares atualizada

Funções auxiliares

isPacMega :: Player -> Bool

Vê se o player está mega

pacMightDie :: Player -> Player

Verifica se o pacman morre ou não

allGhostDead :: [Player] -> [Player]

Torna todos os Ghosts Dead

allGhostAlive :: [Player] -> [Player]

Torna todos os Ghosts Alive

setGhostDead :: Player -> Player

Torna o ghost Dead

setGhostAlive :: Player -> Player

Torna o ghostAlive

setPacMega :: Player -> Player

Torna o pacman Mega

setPlayerPoints :: Player -> Int -> Player

Dado inteiro, soma esse inteiro aos pontos atuais do player

ifAteBigFoodPutAllDead

Arguments

:: [Player]

-> Coords
coordenadas da comida grande

-> Coords
coordenadas do Player

-> Maze

-> [Player]

Verifica se o player comeu a comida grande e retorna lista com os ghosts Dead

isPacmanMega :: [Player] -> Bool

Verifica se existe algum Pacman em modo Mega

spawnMiddle :: Maze -> Player -> Player

Pega num player e coloca-o no meio do maze

respawn :: Maze -> Player -> Player

Respawn do Pacman

changeOrientation :: Orientation -> Int -> [Player] -> [Player]

Muda a orientação de um jogador

Funções finais que juntam as funções de comer e mover do pacman

play :: Play -> State -> State

Função da jogada final e atualiza o maze

updateMaze :: Maze -> Int -> Orientation -> [Player] -> Maze

Atualiza o Maze

moveAuxP

Arguments

:: Maze

-> Int
ID do Player

-> Orientation

-> [Player]

-> [Player]

Movimento tendo em conta a parede e o tunel PACMAN

moveAuxG

Arguments

:: Maze

-> Int
ID do Player

-> Orientation

-> [Player]

-> [Player]

Movimento do Ghost

moveOrTurnPlayerState

Arguments

:: Maze

-> [Player]

-> Player
Player que é movido

-> Orientation

-> [Player]

Muda o estado do Pacman após movimentação e possivelmente comer comida

Função auxiliar

isPacOrGhost :: Player -> Bool

putDead :: Maze -> Player -> [Player] -> Coords -> [Player]

Pega no Player e coloca-o morto ou tira-lhe a vida nas coordenadas no Ghost que o matou

openClose :: Player -> Player

Open Close Mouth

dMega :: Player -> Player

Desce o tempo de mega

Funções auxiliares

findPacman :: [Player] -> Player

Encontra o Pacman numa lista de players

findPlayerID

Arguments

:: [Player]

-> Int
ID do Player

-> Player

Encontra o Player numa lista de Players

findElemInMaze :: Coords -> Maze -> Piece

Encontra um elemento no maze

findNElem :: Int -> [a] -> a

Encontra um elemento no corridor

getPlayerCoordsX :: Player -> Int

Vai buscar a coordenada x dum player

getPlayerCoordsY :: Player -> Int

Vai buscar a coordenada y dum player

Funções auxiliares usadas para fazer testes

printStatsCoords :: State -> [Coords]

printPacMode :: State -> [PacMode]

getGhostMode :: Player -> GhostMode

printGhostMode :: State -> [GhostMode]

getPacTime :: Player -> Double

getPacLife :: Player -> Int

getList :: State -> [Player]

setPlayerVelAux :: Player -> Double -> Player

setPlayerVel :: [Player] -> Int -> Double -> [Player]

printStats :: Player -> String