Tarefa2_2021li1g020

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2021/22.

Documentation

constroiLinha :: [(Peca, Coordenadas)] -> Int -> Int -> Int -> [Peca]

A partir da lista de peças e coordenadas, vai construir a linha constituida pelos blocos, caixas, possível porta e, se o objeto não estiver identificado, e ainda não tiver chegado ao fim da lista, o Vazio.

constroiColunas :: [(Peca, Coordenadas)] -> Int -> Int -> Int -> Mapa

Pegando no constroiLinha vai diminuindo o y (mudar de linha) até chegar ao 0

constroiMapa :: [(Peca, Coordenadas)] -> Mapa

Se o conjunto de Peças e Coordenadas for válido para fazer um mapa, então este pegará no constroiColunas e começando no maior X e maior Y vai diminuido até chegar à posição (0,0). Por esse motivo que se tem de começar adicionando 1 ao obterMaiorX e ao obterMaiorY pois este começa a contar no 0

desconstroiLinha :: [Peca] -> Int -> Int -> [(Peca, Coordenadas)]

O desconstroiLinha pega numa lista de Peças e de acordo com a sua posição adiciona-lhe umas coordenadas e no caso de ser Vazio, não devolve nada

desconstroiColuna :: Mapa -> Int -> [(Peca, Coordenadas)]

O desconstroiColuna vai pegar num Mapa e com o desconstroiLinha vai desconstruir todas as linhas de y==0 ate ao mapa ficar vazio e se tenha uma lista completa de Peças e Coordenadas

desconstroiMapa :: Mapa -> [(Peca, Coordenadas)]

O desconstroiMapa simplesmente pega no desconstroiColuna e começa no y==0