Tarefa3

Description

O objectivo desta tarefa é, dado um labirinto válido, convertê-lo numa sequência de instruções de modo a recriá-lo num formato mais compacto para leitura.

Funções principais

compactMazeHorizontal :: Corridor -> Int -> [(Int, Piece)]

Compactar Corridor

findRepeat

Arguments

:: Instructions

-> Instructions

-> Int
index do primeiro Instructions

-> Int
index do fim da Instructions

-> Instructions

Compara dois Instructions iguais para colocar Repeat

auxFind

Arguments

:: Instruction

-> Instructions

-> Int
index que percorre Instructions

-> Int
index da Instructions de onde foi chamada esta função

-> Instruction

Função auxiliar de findRepeat que procura num Instructions uma Instruction igual à Instruction chamada

compactToInstruction :: Corridor -> Instruction

Compacta o Corridor para Instruction

compactToInstructions :: Maze -> Instructions

Compacta o Maze para Instruction

compactMaze :: Maze -> Instructions

Função principal