Tarefa3_2021li1g020

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2021/22.

Documentation

converterLinha :: [Peca] -> Int -> Int -> Coordenadas -> Direcao -> String

Pega numa lista de Peças, e se os dois inteiros corresponderem às coordenadas então adicionará a personagem, e dependendo da direção será <, no caso de Oeste, ou >, no caso de Este Se for um Bloco adiciona um X, uma Caixa um C, uma Porta um P e se for Vazio simplesmente adiciona um espaço

converterColuna :: Mapa -> Int -> Coordenadas -> Direcao -> String

Usando a função converterLinha, recebe um Mapa e vai transformando cada linha do Mapa numa string, usando depois o "n" para sinalizar que vai mudar de linha

converter :: Jogo -> String

A função converter transforma o Jogo numa String utilizando a converterColuna e adicionando o Jogador na sua posição (x,y)

Orphan instances

Show Jogo
Este show serve para a String conseguir aparecer na consola

Instance details

Methods

showsPrec :: Int -> Jogo -> ShowS

show :: Jogo -> String

showList :: [Jogo] -> ShowS