Tarefa4

Description

Introdução:

O objetivo desta tarefa é fazer movimentar todos os jogadores no mapa, atualizando a Tarefa 2 com funções que permitem a Tarefa4 cumprir o seu objetivo

Objetivos:

Esta tarefa, contém apenas 2 funções:

makeMovesGh -> executa plays, que são calculadas na Tarefa 5, em todos os Ghosts na lista de jogadores

passTime -> função principal que executa a função play no Pacman e junta com as plays do Ghost chamando a função makeMovesGh tomando atenção ao step, de maneira a que faça com que o ghost fique mais lento caso esteja em Modo Dying

Conclusão:

O função corre, no entanto, só considera duas velocidades, quando o step é par ou impar

Documentation

defaultDelayTime :: Integer

Tempo do delay genérico

passTime

Arguments

:: Int
Step

-> State

-> State

Faz play a todos os jogadores

makeMovesGh :: [Play] -> State -> State

Faz as plays do ghosts que são calculadas na tarefa 5