Tarefa4_2021li1g020

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2021/22.

Documentation

cair :: Mapa -> Coordenadas -> Int

A função cair serve para ver quantos y vamos ter de aumentar até o Jogador chegar novamente a um bloco ou caixa ou quando blocos a caixa tem de cair depois de ser largado.

andar :: Jogo -> Movimento -> Jogo

A função andar verifica os movimentos AndarDireita e AndarEsquerda, se estes podem acontecer ou não. AndarDireita: ele não pode andar para a direita se já estiver no ponto extremo direito, se tiver um bloco ou caixa à frente dele e se ele tiver uma caixa e houver um bloco no caminho da caixa. AndarEsquerda: ele não pode andar para a esquerda se já estiver no ponto extremo esquerdo, se tiver um bloco ou caixa à frente dele e se ele tiver uma caixa e houver um bloco no caminho da caixa. Atualmente, como ainda não temos conceito de entrar na Porta, resolvemos fazer com que o jogo fique igual.

trepar :: Jogo -> Jogo

A função trepar serve para o Jogador subir o blococaixa para o qual está a olhar. Se não existir nenhum blococaixa para onde ele está a olhar então a personagem fica no mesmo sítio. Se existirem dois blocos ou duas caixas empilhadas para onde ele está a olhar, ele também não pode trepar, visto que o Jogador só consegue trepar um bloco. Se ele tiver uma caixa e quando o Jogador trepar, a caixa ficasse no local onde existe um bloco, então ele já não pode trepar.

manuziarCaixa :: Mapa -> Coordenadas -> Peca -> Mapa

O manuziarCaixa simplesmente vê onde é que a caixa fica ou o que tem de ficar vazio depois de se ter interagido com ela.

interageCaixa :: Jogo -> Jogo

O interageCaixa pega no manuziarCaixa e vê todos os casos possíveis: quando a personagem já tem caixa, quando consegue pegar nela, se ele está a olhar para ela e assim consegue pegar nela, ou se pode ser colocado para o sítio onde estão a olhar, ou quando blocos ela tem de cair depois de ser largada.

moveJogador :: Jogo -> Movimento -> Jogo

O moveJogador simplesmente pega em todos os casos de movimento e interpreta-os de acordo com a sua função.

correrMovimentos :: Jogo -> [Movimento] -> Jogo

O correrMovimentos simplesmente pega no moveJogador e utiliza uma lista de movimentos que vê da esquerda para a direita.