Tarefa5

Description

Introdução:

O objetivo desta tarefa é implementar um comportamento para os fantasmas.

Objetivos:

Esta tarefa, contém 3 funções principais:

chaseMode -> retorna jogadas que permitem aos Ghosts perseguirem o Pacman

scatterMode -> retorna jogadas que permitem aos Ghosts fugirem do Pacman

ghostPlay -> Junta as funções scatterMode e chaseMode, chamando cada uma na situação mais pertinente

Estratégias:

chaseMode : nesta função a estratégia adotada foi verificar as coordenadas x e y do Pacman e fazer com que as coordenadas x e y do Ghost aumentem ou diminuem até serem iguais ao Pacman, tendo atenção às Walls de maneira a que não fique preso na Wall. Quando o Ghost é comido, esse ghost executa duas Plays para cima de maneira a sair da GhostHouse e depois prossegue com o seu comportamento de perseguição.

scatterMode : nesta função a estratégia adotada foi verificar as coordenadas x e y do Pacman e fazer com que as coordenadas x e y do Ghost aumentem ou diminuem, de maneira a que haja maior amplitude entre as coordenadas entre o Pacman e Ghost, também é verificado caso o Ghost está em algum canto do mapa, se mover para o lado ou comprimento oposto do Pacman

Conclusão:

Os Ghosts em chaseMode, de facto, perseguem o Pacman, no entando o scatterMode, fica "confuso" quando se encontra nos cantos.

Função principal

ghostPlay :: State -> [Play]

Jogadas do Ghost

Função ChaseMode

chaseMode

Arguments

:: State

-> Int
ID do Ghost

-> Play

Perseguir o Pacman

whereToFollow

Arguments

:: Int
ID do Ghost

-> [Player]

-> Coords
Coordenadas do Pacman

-> Orientation

Devolve orientação em direção à posição do Pacman

stillInHouse

Arguments

:: State

-> Int
ID do Ghost

-> Bool

Condições de verificação se o Ghost ainda está dentro da casa

checkWallsToFollow

Arguments

:: State

-> Int
ID do Ghost

-> Coords
Coordenadas do Pacman

-> Orientation

Verifica condições caso esteja contra a parede

Função Scatter Mode

scatterMode :: State -> Int -> Play

Fugir do Pacman

whereToRun :: Int -> [Player] -> Coords -> Orientation

Devolve orientação contrária à posição do Pacman

checkWallsToRun

Arguments

:: State

-> Int
ID do Ghost

-> Coords
Coordenadas do Pacman

-> Orientation

Verifica condições caso esteja contra a parede ou nos cantos

runAway

Arguments

:: Int
ID do Pacman

-> State

-> Play

Função auxiliar da scatter Mode que contem a play